feat: filling out more of the File API, making optional
