+++ src/scikit_build_core/cmake.py
@@ -13,7 +13,6 @@ from packaging.version import Version
 from ._logging import logger
 from ._shutil import Run
 from .errors import CMakeAccessError, CMakeConfigError, FailedLiveProcessError
-from .file_api import Index, get_index
 
 __all__ = ["CMake", "CMakeConfig", "get_cmake_path"]
 
@@ -129,20 +128,11 @@ class CMakeConfig:
 
         _cmake_args += cmake_args or []
 
-        query_dir = self.api_dir / "query"
-        query_dir.mkdir(parents=True)
-        query_dir.joinpath("codemodel-v2").touch()
-
         try:
             Run().live(self.cmake, *_cmake_args)
         except subprocess.CalledProcessError:
             msg = "CMake configuration failed"
             raise FailedLiveProcessError(msg) from None
-
-    def get_index(self) -> Index:
-        reply_dir = self.api_dir / "reply"
-        assert reply_dir.is_dir(), "Must be run after .configure()"
-        return get_index(reply_dir)
 
     def build(self) -> None:
         opts: dict[str, str] = {}
