+++ src/scikit_build_core/file_api/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
