+++ src/scikit_build_core/file_api/model/common.py
@@ -1,0 +1,14 @@
+import dataclasses
+from pathlib import Path
+
+
+@dataclasses.dataclass
+class APIVersion:
+    major: int
+    minor: int
+
+
+@dataclasses.dataclass
+class Paths:
+    source: Path
+    build: Path
